using Hayden's modularization code and testing functionality, update and faqs command working locally
