+++ faqsDB.json
@@ -19,6 +19,11 @@
       "keyword": "test",
       "question": "is this actually updating our DB ?",
       "answer": "Only one way to find out"
+    },
+    {
+      "keyword": "testTwo",
+      "question": "trying to update db",
+      "answer": "will I work today?"
     }
   ]
 }
