--- labDB.json
@@ -1,14 +1,0 @@
-{
-  "data": [
-    {
-      "assignment": "Lab 01:",
-      "instructions": "Learn the basics of creating a back end server",
-      "submission": "Submit a link to your completed README.md."
-    },
-    {
-      "assignment": "Lab 02:",
-      "instructions": "Create a express server able to get information from a database",
-      "submission": "Submit a link to your completed README.md."
-    }
-  ]
-}
