+++ package.json
@@ -2,7 +2,7 @@
   "name": "chatbotbackend",
   "version": "1.0.0",
   "description": "## Authors: Donna Ada, Kao Saelor, Kenya Womack, Hayden Cooper, Tricia Sawyer, Reece Renninger",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
